Move base model to shared/model and adjust path to shared
This changes shared include path so it points to "shared/src"

Which will allow consumers of shared modules to easily just include the file
so we don't have to add everything to the shared index
