--- shared/index.cjsx
@@ -1,62 +1,0 @@
-{PinnedHeader, CardBody, PinnableFooter} = require './src/components/pinned-header-footer-card/sections'
-{Exercise, ExerciseWithScroll} = require './src/components/exercise'
-{ExControlButtons} = require './src/components/exercise/controls'
-{APIHandler, APIHandlerBase} = require './src/api'
-
-module.exports = {
-
-  OXRouter:               require './src/helpers/router'
-  OXLink:                 require './src/factories/link'
-  OXButtonLink:           require './src/factories/button-link'
-  OXMatchByRouter:        require './src/helpers/match-by-router'
-
-  APIHandler, APIHandlerBase,
-  APIActionAdapter:       require './src/api/action-adapter'
-  ReactHelpers:           require './src/helpers/react'
-
-  ArbitraryHtmlAndMath:   require './src/components/html'
-  Markdown:               require './src/components/markdown'
-  AsyncButton:            require './src/components/buttons/async-button'
-  BootstrapURLs:          require './src/model/urls'
-  ChangeStudentIdForm:    require './src/components/change-student-id-form'
-  UiSettings:             require './src/model/ui-settings'
-  Networking:             require './src/model/networking'
-  Breadcrumb:             require './src/components/breadcrumb'
-  CardBody,
-  ChapterSectionMixin:    require './src/components/chapter-section-mixin'
-  CloseButton:            require './src/components/buttons/close-button'
-  CcJoinConflict:         require './src/components/enroll/cc-join-conflict'
-  ConfirmJoinCourse:      require './src/components/enroll/confirm-join-course'
-  MessageList:            require './src/components/enroll/message-list'
-
-  Exercise,
-  ExControlButtons,
-  ExerciseGroup:          require './src/components/exercise/group'
-  ExerciseIdentifierLink: require './src/components/exercise-identifier-link'
-  ExerciseHelpers:        require './src/model/exercise'
-  ExercisePreview:        require './src/components/exercise-preview'
-  ExerciseWithScroll,
-  TaskHelper:             require './src/helpers/task'
-  StepHelpsHelper:        require './src/helpers/step-helps'
-  propHelpers:            require './src/helpers/props'
-  Logging:                require './src/helpers/logging'
-  ExerciseIntro:          require './src/components/exercise/intro'
-
-  FreeResponse:           require './src/components/exercise/free-response'
-  GetPositionMixin:       require './src/components/get-position-mixin'
-  ScrollToMixin:          require './src/components/scroll-to-mixin'
-  KeysHelper:             require './src/helpers/keys'
-  NotificationActions:    require './src/model/notifications'
-  NotificationsBar:       require './src/components/notifications/bar'
-  PinnableFooter,
-  PinnedHeader,
-  PinnedHeaderFooterCard: require './src/components/pinned-header-footer-card'
-  Question:               require './src/components/question'
-  RefreshButton:          require './src/components/buttons/refresh-button'
-  ResizeListenerMixin:    require './src/components/resize-listener-mixin'
-  SmartOverflow:          require './src/components/smart-overflow'
-  SpyMode:                require './src/components/spy-mode'
-  SuretyGuard:            require './src/components/surety-guard'
-  HandleBodyClassesMixin: require './src/components/handle-body-classes-mixin'
-
-}
