+++ shared/src/helpers/lazy-getter.js
