+++ tutor/resources/styles/tutor.scss
@@ -1,5 +1,5 @@
 // shared openstax components
-@import '~shared/resources/styles/main';
+@import '../../../shared/resources/styles/main';
 
 // Variables and Mixins
 // TODO to many renamed variables running around.
