+++ tutor/specs/models/course.spec.js
@@ -1,4 +1,4 @@
-import UiSettings from 'shared/src/model/ui-settings';
+import UiSettings from 'shared/model/ui-settings';
 import { map, cloneDeep, shuffle } from 'lodash';
 import Courses from '../../src/models/courses-map';
 import Course from '../../src/models/course';
@@ -9,7 +9,7 @@ import { bootstrapCoursesList } from '../courses-test-data';
 
 import COURSE from '../../api/courses/1.json';
 
-jest.mock('shared/src/model/ui-settings');
+jest.mock('shared/model/ui-settings');
 
 describe('Course Model', () => {
