+++ tutor/specs/models/jobs/scores-export.spec.js
@@ -1,4 +1,4 @@
-import UiSettings from 'shared/src/model/ui-settings';
+import UiSettings from 'shared/model/ui-settings';
 import Export from '../../../src/models/jobs/scores-export';
 import { bootstrapCoursesList } from '../../courses-test-data';
 import { Completed } from '../../../src/models/jobs/queue';
@@ -9,7 +9,7 @@ jest.mock('../../../src/flux/time', () => ({
   TimeStore: { getNow: jest.fn(() => mockNowDate) },
 }));
 
-jest.mock('shared/src/model/ui-settings');
+jest.mock('shared/model/ui-settings');
 
 describe('Scores export job', () => {
