+++ tutor/src/components/icons/best-practices.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import StaticComponent from 'shared/src/components/static';
+import StaticComponent from 'shared/components/static';
 
 export default class BestPracticesIcon extends StaticComponent {
   render() {
